requirements.txt
aio-pika
anthropic
python-dotenv